refactor: move gamelog to new module and don't export macros
