+++ src/errors.rs
@@ -109,7 +109,6 @@ mod macros {
         $crate::errors::PoksError::internal(format!($($toks)+))
     };
 }
-
     pub(crate) use err_int;
 }
 pub(crate) use macros::*;
